NTBS-207: Move Notification Error generation into helper class
